+++ packages/runtime/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./runtime-dsl-parser";
+export * from "./dependency-graph";
 export * from "./template-resolver";
 export * from "./types";